CA-47596: Strip VDI.snapshot decoration.
Since SMs are presently supposed to deal with the tap-level pause
themselves, this left nothing but a gratuitous rename from SR.snapshot
to SR.snapshot_paused.

Removed, in favor of having to alias .snapshot_paused to .snasphot in
the base VDI class. Preferable for SR types which carry tapdisks but
don't need a software pause, such as iSCSI filers with hardware
snapshots since CA-46683.

Signed-off-by: Daniel Stodden <daniel.stodden@citrix.com>