SCTX-759: enable recovery of an unplugged SR during SR.attach.
Interrupted leaf-coalesce operations are handled by the GC module, which
expects the SR to be attached. If the PBD is not plugged and the SR is
in an interrupted-leaf-coalesce state, PBD.plug will fail because SR.attach
will attempt to perform recovery by calling the GC module, which in turn will
fail because the SR is not yet attached. We break this cycle by allowing the
GC module to proceed in this case even if the SR is not attached.